requirements.txt
mcp[cli]>=1.0.0
llama-cloud>=0.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
starlette>=0.36.0
